+++ packages/frontend/client/src/utils.ts
@@ -2,7 +2,7 @@
 export const setImmediate = (handler: () => any) => setTimeout(handler, 0);
 
 export const rateLimit = <T extends (...args: any[]) => void>(func: T, delay: number) => {
-    let timeout: NodeJS.Timeout | undefined = undefined;
+    let timeout: number | undefined = undefined;
     let lastTime = 0;
 
     return function(...args: Parameters<T>) {
@@ -17,7 +17,7 @@ export const rateLimit = <T extends (...args: any[]) => void>(func: T, delay: nu
             return;
         } else {
             // Not timeout exists and not enough time has passed, set a timeout
-            timeout = setTimeout(() => {
+            timeout = window.setTimeout(() => {
                 timeout = undefined;
                 lastTime = Date.now();
                 func(...args);